+++ projeto-obra/src/main/java/com/projeto/obra/entities/ObraDetalhesTecnicos.java
@@ -1,7 +1,5 @@
 package com.projeto.obra.entities;
 
-import java.math.BigInteger;
-
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
@@ -15,7 +13,7 @@ public class ObraDetalhesTecnicos {
     
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private BigInteger id;
+    private Integer id;
     @ManyToOne
     private Obra obraID;
     @Enumerated(EnumType.STRING)
@@ -26,11 +24,11 @@ public class ObraDetalhesTecnicos {
     public ObraDetalhesTecnicos() {
     }
 
-    public BigInteger getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Integer id) {
         this.id = id;
     }
